去除io多线程操作，去除request多线程操作
