+++ src/main/workers/utils/index.ts
@@ -1,1 +1,1 @@
-export * as Io from './io';
+export * as Coding from './coding';
